service 수정, 인증 uri 설정

+++ src/main/java/com/fhdufhdu/mim/entity/PostingId.java
@@ -23,5 +23,5 @@ public class PostingId implements Serializable {
     private MovieBoard movieBoard;
 
     @NotNull
-    private Integer commentCnt;
+    private Long commentCnt;
 }
